more correctly record log changes after events w/ juggler w/ 1 achievement
The problem is that when adding the first achievement for a juggler or removing
the last achievement the total number of jugglers we take in to account for all
achievements is altered.  So, when these events happen we need to record a
changelog entry for every achievement (where > 1 juggler has achieved it) and
every juggler (with > 0 achievements).  We do not need to record logs for achievements
with only 1 juggler because their value is already 100 and will not change when the
total number of jugglers changes.  Ideally, we would only write log entries for jugglers
with at least one achievement that has also been achieved by someone else.  If all of
the achievements that a juggler has have only been achieved by them, then their score
will not be affected when a new juggler comes along.  This has not been implemented yet.